requirements.txt
fastapi
uvicorn
tflite-runtime==2.14.0
pillow
numpy>=1.24.0,<2.0
python-multipart
